moved layout inside components and removed duplicated code
