+++ sources/packages/web/src/components/layouts/Institution/sidebar/HomeSideBar.vue
@@ -1,5 +1,5 @@
 <template>
-  <Menu :model="items"/>
+  <Menu :model="items" />
 </template>
 <script lang="ts">
 import Menu from "primevue/menu";
@@ -39,7 +39,7 @@ export default {
     ]);
 
     return {
-      items
+      items,
     };
   },
 };
