+++ sources/packages/web/src/views/institution/Institution.vue
@@ -5,11 +5,11 @@
 </template>
 
 <script lang="ts">
-import InstitutionFullLayout from '../../layouts/InstitutionFullLayout.vue'
+import InstitutionFullLayout from "../../components/layouts/Institution/InstitutionFullLayout.vue";
 
 export default {
   components: {
     InstitutionFullLayout,
-  }, 
+  },
 };
 </script>
